Install libsqlite3-dev for Zulip test on Linux
This package is no longer included in GitHub's new version of the
"ubuntu-latest" images.  That causes nondeterministic failures when
the runner happens to use the new version:
  https://github.com/flutter/tests/pull/441#issuecomment-2593993199

The actual shared library file libsqlite3.so.0.8.6 -- in the package
libsqlite3-0 -- is still installed by default anyway.  This package
adds the symlink at "libsqlite3.so", which is the filename some
libraries apparently look for.