Update Zulip tests and re-enable (#441)
This has the update to unblock autocomplete changes:
  https://github.com/zulip/zulip-flutter/pull/1190
  https://github.com/flutter/flutter/pull/143249

and with that, tests pass again on Flutter main.

Also install libsqlite3-dev on Linux, a package which is no longer
included in GitHub's new version of the "ubuntu-latest" images.  The
lack of it causes nondeterministic failures when the runner happens
to use the new image:
  https://github.com/flutter/tests/pull/441#issuecomment-2593993199

The actual shared library file libsqlite3.so.0.8.6 -- in the package
libsqlite3-0 -- is still installed by default anyway.  This package
adds the symlink at "libsqlite3.so", which is the filename some
libraries apparently look for.